refactor object store
